requirements.txt
jsonpath-ng
atlassian-python-api
requests
db-sqlite3
business_calendar
rule_engine
boto3
slack_sdk
pytz
psycopg2-binary
pandas
fastapi>=0.68.0
pydantic<2
email-validator>=2.0.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
